feat: +Shell01 +C04

+++ C07/ex04/ft_convert_base.c
@@ -43,7 +43,7 @@ char	*ft_totoa(char *str, char *base)
 	return (itoa);
 }
 
-int	ft_convert_dec_base(char *nbr, char *base)
+long	ft_convert_dec_base(char *nbr, char *base)
 {
 	int		i;
 	int		len;
